requirements.txt
flask
flask_cors
psycopg2
psycopg2-binary
python-dotenv
bcrypt
sqlalchemy
apscheduler